Reducer Files reorganized
* Categorized reducers file into different files
* Updated logic for merge of event lists for Eventlogs in reducer
