+++ frontend/.eslintrc.js
@@ -37,10 +37,7 @@ module.exports = {
         "import/no-unresolved": 0,
         'react/jsx-props-no-spreading': 0,
         "react/require-default-props": 0,
-        "no-use-before-define": [
-          "error",
-          { "functions": false, "classes": true, "variables": false }
-        ],
+        "no-use-before-define": "off",
         "import/no-extraneous-dependencies": ["error", {"devDependencies": true}],
         "prettier/prettier": "error",
         "@typescript-eslint/explicit-function-return-type": "off",
