+++ frontend/src/store/controller/reducers.ts
@@ -4,392 +4,22 @@ import {
   SensorMeasurements,
   CycleMeasurements,
   Parameters,
-  ParametersRequest,
   Ping,
   Announcement,
-  AlarmLimitsRequest,
-  VentilationMode,
   ActiveLogEvents,
-  ExpectedLogEvent,
-  LogEvent,
-  NextLogEvents,
 } from './proto/mcu_pb';
+import { MessageType } from './types';
+import { waveformHistoryReducer, pvHistoryReducer } from './reducers/derived';
 import {
-  RotaryEncoder,
-  FrontendDisplaySetting,
-  ThemeVariant,
-  Unit,
-  SystemSettingRequest,
-} from './proto/frontend_pb';
-import {
-  MessageType,
-  ParameterCommitAction,
-  PARAMETER_COMMITTED,
-  PBMessage,
-  PBMessageType,
-  StateUpdateAction,
-  STATE_UPDATED,
-  WaveformHistory,
-  WaveformPoint,
-  PVHistory,
-  PVPoint,
-  ALARM_LIMITS,
-  FRONTEND_DISPLAY_SETTINGS,
-  SYSTEM_SETTINGS,
-  commitAction,
-  EXPECTED_LOG_EVENT_ID,
-  RotaryEncoderParameter,
-  ALARM_LIMITS_STANDBY,
-  PARAMETER_STANDBY,
-} from './types';
-import DECIMAL_RADIX from '../../modules/app/AppConstants';
-
-const messageReducer = <T extends PBMessage>(
-  messageType: MessageType,
-  pbMessageType: PBMessageType,
-) => (state: T = pbMessageType.fromJSON({}) as T, action: StateUpdateAction): T => {
-  switch (action.type) {
-    case STATE_UPDATED:
-      if (action.messageType === messageType) {
-        if (messageType === MessageType.NextLogEvents) {
-          return mergeEventList(action.state as NextLogEvents, state as NextLogEvents) as T;
-        }
-        if (action.messageType === MessageType.RotaryEncoder) {
-          return calculateStepDiff(action.state as RotaryEncoder, state as RotaryEncoder) as T;
-        }
-        return action.state as T;
-      }
-      return state;
-    default:
-      return state;
-  }
-};
-
-const mergeEventList = (newEvents: NextLogEvents, oldEvents: NextLogEvents): NextLogEvents => {
-  const events = [...oldEvents.logEvents];
-  if (!oldEvents || !oldEvents.logEvents.length) {
-    return newEvents as NextLogEvents;
-  }
-  newEvents.logEvents.forEach((logEvent: LogEvent) => {
-    const id = events.find((ev: LogEvent) => ev.id === logEvent.id);
-    if (id === undefined) {
-      events.push(logEvent);
-    }
-  });
-  return {
-    ...newEvents,
-    logEvents: events,
-  };
-};
-
-const calculateStepDiff = (
-  newState: RotaryEncoder,
-  oldState: RotaryEncoder,
-): RotaryEncoderParameter => {
-  const stepDiff = newState.step - oldState.step;
-  const stateCopy = { ...newState } as RotaryEncoderParameter;
-  stateCopy.stepDiff = stepDiff;
-  return stateCopy;
-};
-
-const alarmLimitsReducer = (
-  state: AlarmLimitsRequest = AlarmLimitsRequest.fromJSON({
-    rrMax: 100,
-    pipMax: 100,
-    peepMax: 100,
-    ipAbovePeepMax: 100,
-    inspTimeMax: 100,
-    fio2Max: 100,
-    pawMax: 100,
-    mveMax: 100,
-    tvMax: 100,
-    etco2Max: 100,
-    flowMax: 100,
-    apneaMax: 100,
-    spo2Max: 100,
-  }) as AlarmLimitsRequest,
-  action: commitAction,
-): AlarmLimitsRequest => {
-  return withRequestUpdate<AlarmLimitsRequest>(state, action, ALARM_LIMITS);
-};
-
-const alarmLimitsRequestStandbyReducer = (
-  state: { alarmLimits: AlarmLimitsRequest } = {
-    alarmLimits: AlarmLimitsRequest.fromJSON({
-      rrMax: 100,
-      pipMax: 100,
-      peepMax: 100,
-      ipAbovePeepMax: 100,
-      inspTimeMax: 100,
-      fio2Max: 100,
-      pawMax: 100,
-      mveMax: 100,
-      tvMax: 100,
-      etco2Max: 100,
-      flowMax: 100,
-      apneaMax: 100,
-      spo2Max: 100,
-    }),
-  } as { alarmLimits: AlarmLimitsRequest },
-  action: commitAction,
-): { alarmLimits: AlarmLimitsRequest } => {
-  switch (action.type) {
-    case STATE_UPDATED: // ignore message from backend
-      return state;
-    case `@controller/${ALARM_LIMITS_STANDBY}_COMMITTED`:
-      return {
-        alarmLimits: {
-          ...state.alarmLimits,
-          ...action.update,
-        } as AlarmLimitsRequest,
-      };
-    default:
-      return state;
-  }
-};
-
-const parametersRequestStanbyReducer = (
-  state: { parameters: ParametersRequest } = {
-    parameters: ParametersRequest.fromJSON({
-      mode: VentilationMode.hfnc,
-      pip: 30,
-      peep: 0,
-      rr: 30,
-      ie: 1.0,
-      fio2: 60.0,
-    }),
-  } as { parameters: ParametersRequest },
-  action: commitAction,
-): { parameters: ParametersRequest } => {
-  switch (action.type) {
-    case STATE_UPDATED: // ignore message from backend
-      return state;
-    case `@controller/${PARAMETER_STANDBY}_COMMITTED`:
-      return {
-        parameters: {
-          ...state.parameters,
-          ...action.update,
-        } as ParametersRequest,
-      };
-    default:
-      return state;
-  }
-};
-
-const frontendDisplaySettingReducer = (
-  state: FrontendDisplaySetting = FrontendDisplaySetting.fromJSON({
-    theme: ThemeVariant.dark,
-    unit: Unit.imperial,
-  }) as FrontendDisplaySetting,
-  action: commitAction,
-): FrontendDisplaySetting => {
-  return withRequestUpdate<FrontendDisplaySetting>(state, action, FRONTEND_DISPLAY_SETTINGS);
-};
-
-const expectedLoggedEventReducer = (
-  state: ExpectedLogEvent = ExpectedLogEvent.fromJSON({
-    id: 0,
-  }) as ExpectedLogEvent,
-  action: commitAction,
-): ExpectedLogEvent => {
-  return withRequestUpdate<ExpectedLogEvent>(state, action, EXPECTED_LOG_EVENT_ID);
-};
-
-const systemSettingRequestReducer = (
-  state: SystemSettingRequest = SystemSettingRequest.fromJSON({
-    brightness: 100,
-    date: parseInt((new Date().getTime() / 1000).toFixed(0), DECIMAL_RADIX),
-  }) as SystemSettingRequest,
-  action: commitAction,
-): SystemSettingRequest => {
-  return withRequestUpdate<SystemSettingRequest>(state, action, SYSTEM_SETTINGS);
-};
-
-const withRequestUpdate = <T>(state: T, action: commitAction, prefix: string): T => {
-  switch (action.type) {
-    case STATE_UPDATED: // ignore message from backend
-      return state;
-    case `@controller/${prefix}_COMMITTED`:
-      return { ...state, ...action.update } as T;
-    default:
-      return state;
-  }
-};
-
-const parametersRequestReducer = (
-  state: ParametersRequest = ParametersRequest.fromJSON({
-    mode: VentilationMode.hfnc,
-    pip: 30,
-    peep: 0,
-    rr: 30,
-    ie: 1.0,
-    fio2: 60.0,
-  }) as ParametersRequest,
-  action: StateUpdateAction | ParameterCommitAction,
-): ParametersRequest => {
-  switch (action.type) {
-    case STATE_UPDATED: // ignore message from backend
-      return state;
-    case PARAMETER_COMMITTED:
-      return { ...state, ...action.update };
-    default:
-      return state;
-  }
-};
-
-const waveformHistoryReducer = <T extends PBMessage>(
-  messageType: MessageType,
-  getTime: (values: T) => number,
-  getValue: (values: T) => number,
-  bufferDuration = 60,
-  segmentUpdateOffset = 0,
-  maxDuration = 10000,
-  gapDuration = 500,
-  maxSegmentDuration = 2500,
-) => (
-  state: WaveformHistory = {
-    waveformOld: {
-      full: [],
-    },
-    waveformNew: {
-      full: [],
-      buffer: [],
-      segmented: [[]],
-    },
-    waveformNewStart: 0,
-  },
-  action: StateUpdateAction,
-): WaveformHistory => {
-  switch (action.type) {
-    case STATE_UPDATED:
-      if (action.messageType === messageType) {
-        const sampleTime = getTime(action.state as T);
-        const lastTime =
-          state.waveformNew.full.length === 0
-            ? state.waveformNewStart
-            : state.waveformNew.full[state.waveformNew.full.length - 1].date;
-        if (
-          sampleTime > state.waveformNewStart + maxDuration ||
-          sampleTime < lastTime ||
-          new Date(sampleTime - gapDuration - state.waveformNewStart) > lastTime
-        ) {
-          // make waveformNew start over
-          const newPoint = {
-            date: new Date(0),
-            value: getValue(action.state as T),
-          };
-          return {
-            waveformOld: {
-              full: state.waveformNew.full,
-            },
-            waveformNew: {
-              full: [],
-              buffer: [newPoint],
-              segmented: [[]],
-            },
-            waveformNewStart: sampleTime,
-          };
-        }
-
-        // update buffer
-        let buffered = [...state.waveformNew.buffer];
-        const newPointTime = sampleTime - state.waveformNewStart;
-        const newPoint = {
-          date: new Date(newPointTime),
-          value: getValue(action.state as T),
-        };
-        buffered = buffered.concat([newPoint]);
-
-        // apply segment update offset
-        let segments = [...state.waveformNew.segmented];
-        const bufferedStart = buffered[0].date.getTime();
-        const bufferedEnd = buffered[buffered.length - 1].date.getTime();
-        if (segments.length === 1 && bufferedEnd - bufferedStart < segmentUpdateOffset) {
-          return {
-            ...state,
-            waveformNew: {
-              ...state.waveformNew,
-              buffer: buffered,
-            },
-          };
-        }
-
-        // update segmented
-        const lastSegment = segments[segments.length - 1];
-        if (lastSegment.length === 0) {
-          segments[segments.length - 1] = buffered;
-          buffered = [];
-        } else {
-          const lastSegmentStart = lastSegment[0].date.getTime();
-          const lastSegmentEnd = lastSegment[lastSegment.length - 1].date.getTime();
-          if (newPointTime - lastSegmentEnd >= bufferDuration) {
-            const lastSegmentDuration = newPointTime - lastSegmentStart;
-            if (lastSegmentDuration >= maxSegmentDuration) {
-              // start a new segment, but add an overlap of points
-              segments[segments.length - 1] = lastSegment.concat([buffered[0]]);
-              segments = segments.concat([buffered]);
-            } else {
-              segments[segments.length - 1] = lastSegment.concat(buffered);
-            }
-            buffered = [];
-          }
-        }
-
-        return {
-          ...state,
-          waveformNew: {
-            full: state.waveformNew.full.concat([newPoint]),
-            buffer: buffered,
-            segmented: segments,
-          },
-        };
-      }
-      return state;
-    default:
-      return state;
-  }
-};
-
-const pvHistoryReducer = (
-  state: PVHistory = {
-    loop: [],
-    loopOrigin: {
-      pressure: 0,
-      volume: 0,
-    },
-    cycle: 0,
-  },
-  action: StateUpdateAction,
-): PVHistory => {
-  switch (action.type) {
-    case STATE_UPDATED:
-      if (action.messageType === MessageType.SensorMeasurements) {
-        const newState = action.state as SensorMeasurements;
-        const { cycle, paw, volume } = newState;
-        if (cycle !== state.cycle) {
-          // make loop start over
-          return {
-            loop: [{ pressure: 0, volume: 0 }],
-            loopOrigin: { pressure: paw, volume },
-            cycle,
-          };
-        }
-
-        // add point to loop
-        const newPoint = {
-          pressure: paw - state.loopOrigin.pressure,
-          volume: volume - state.loopOrigin.volume,
-        };
-        return {
-          ...state,
-          loop: state.loop.concat([newPoint]),
-        };
-      }
-      return state;
-    default:
-      return state;
-  }
-};
+  alarmLimitsReducer,
+  alarmLimitsRequestStandbyReducer,
+  expectedLoggedEventReducer,
+  frontendDisplaySettingReducer,
+  parametersRequestReducer,
+  parametersRequestStanbyReducer,
+  systemSettingRequestReducer,
+} from './reducers/components';
+import { messageReducer, nextLogEventsReducer, rotaryEncoderReducer } from './reducers/backend';
 
 export const controllerReducer = combineReducers({
   // Message states from mcu_pb
@@ -400,7 +30,7 @@ export const controllerReducer = combineReducers({
   systemSettingRequest: systemSettingRequestReducer,
   frontendDisplaySetting: frontendDisplaySettingReducer,
   expectedLoggedEvent: expectedLoggedEventReducer,
-  nextLogEvents: messageReducer<NextLogEvents>(MessageType.NextLogEvents, NextLogEvents),
+  nextLogEvents: nextLogEventsReducer,
   activeLogEvents: messageReducer<ActiveLogEvents>(MessageType.ActiveLogEvents, ActiveLogEvents),
   sensorMeasurements: messageReducer<SensorMeasurements>(
     MessageType.SensorMeasurements,
@@ -416,7 +46,7 @@ export const controllerReducer = combineReducers({
   announcement: messageReducer<Announcement>(MessageType.Announcement, Announcement),
 
   // Message states from frontend_pb
-  rotaryEncoder: messageReducer<RotaryEncoderParameter>(MessageType.RotaryEncoder, RotaryEncoder),
+  rotaryEncoder: rotaryEncoderReducer,
 
   // Derived states
   waveformHistoryPaw: waveformHistoryReducer<SensorMeasurements>(
